+++ application/src/main/java/common/param/sl/DepositEvent.java
@@ -38,7 +38,7 @@ public class DepositEvent extends TxnEvent {
      * @param accountTransfer
      * @param bookEntryTransfer
      */
-    public DepositEvent(int bid, int pid, String bid_array, String partition_index, int num_of_partition,
+    public DepositEvent(double bid, int pid, String bid_array, String partition_index, int num_of_partition,
                         String accountId,
                         String bookEntryId,
                         long accountTransfer,
@@ -88,6 +88,6 @@ public class DepositEvent extends TxnEvent {
     }
 
     public DepositEvent cloneEvent() {
-        return new DepositEvent((int) bid, pid, Arrays.toString(bid_array), Arrays.toString(partition_indexs), number_of_partitions, accountId, bookEntryId, accountTransfer, bookEntryTransfer);
+        return new DepositEvent(bid, pid, Arrays.toString(bid_array), Arrays.toString(partition_indexs), number_of_partitions, accountId, bookEntryId, accountTransfer, bookEntryTransfer);
     }
 }
